[ios] Rename private header.
