+++ platform/ios/src/MGLApplication_Private.h
